sparsification

Laplacians.approxQual

Laplacians.conditionNumber

Laplacians.conditionNumber

Laplacians.sparsify

Laplacians.support

Laplacians.sparsify — Method

as = sparsify(a; ep=0.5)

Apply Spielman-Srivastava sparsification: sampling by effective resistances. ep should be less than 1.

Laplacians.approxQual — Method

eps = approxQual(graph1, graph2; tol=1e-5, verbose=false)

Computes the eps for which graph1 and graph2 are eps approximations of each other. That is, L1 <= (1+eps) L2, and vice versa.

It is randomized, so you might want to run it again if you don't trust the answers.

Laplacians.conditionNumber — Method

kappa = conditionNumber(graph, precon; tol=1e-5, verbose=false)

Computes the relative condition number of graph and a preconditioning function.

It is randomized, so you might want to run it again if you don't trust the answers.

Laplacians.conditionNumber — Method

kapps = conditionNumber(graph1, graph2; tol=1e-5, verbose=false)

Computes the relative condition number of graph1 and graph2.

It is randomized, so you might want to run it again if you don't trust the answers.

Laplacians.support — Method

sup12, sup21 = support(graph1, graph2; tol=1e-5)

Computes the support of graph1 wrt graph2, and the other way around. It is randomized, so you might want to run it again if you don't trust the answers.