akpw

AKPW

Also see the page on Low Stretch Spanning Trees

Laplacians.akpw

Laplacians.akpwU

Laplacians.akpw — Method.

tree = akpw(graph; ver=0)

Computes a low stretch spanning tree of graph, and returns it as a graph. The default version is 0. In event of emergency, one can try ver=2. It is usually slower, but might have slightly better stretch.

source

Laplacians.akpwU — Method.

tree = akpwU(graph)

Computes a low stretch spanning tree of an unweighted graph, and returns it as a graph.

source